Remote value converter couple
